+++ src/basic/__tests__/basic.test.tsx
@@ -1,11 +1,18 @@
-import { useState } from "react";
+import { useState } from 'react';
 import { describe, expect, test } from 'vitest';
-import { act, fireEvent, render, renderHook, screen, within } from '@testing-library/react';
-import { CartPage } from '../../refactoring/components/CartPage';
-import { AdminPage } from "../../refactoring/components/AdminPage";
+import {
+  act,
+  fireEvent,
+  render,
+  renderHook,
+  screen,
+  within,
+} from '@testing-library/react';
+import { CartPage } from '../../refactoring/components/cart/CartPage';
+import { AdminPage } from '../../refactoring/components/AdminPage';
 import { CartItem, Coupon, Product } from '../../types';
-import { useCart, useCoupons, useProducts } from "../../refactoring/hooks";
-import * as cartUtils from "../../refactoring/hooks/utils/cartUtils";
+import { useCart, useCoupons, useProducts } from '../../refactoring/hooks';
+import * as cartUtils from '../../refactoring/hooks/utils/cartUtils';
 
 const mockProducts: Product[] = [
   {
@@ -13,55 +20,54 @@ const mockProducts: Product[] = [
     name: '상품1',
     price: 10000,
     stock: 20,
-    discounts: [{ quantity: 10, rate: 0.1 }]
+    discounts: [{ quantity: 10, rate: 0.1 }],
   },
   {
     id: 'p2',
     name: '상품2',
     price: 20000,
     stock: 20,
-    discounts: [{ quantity: 10, rate: 0.15 }]
+    discounts: [{ quantity: 10, rate: 0.15 }],
   },
   {
     id: 'p3',
     name: '상품3',
     price: 30000,
     stock: 20,
-    discounts: [{ quantity: 10, rate: 0.2 }]
-  }
+    discounts: [{ quantity: 10, rate: 0.2 }],
+  },
 ];
 const mockCoupons: Coupon[] = [
   {
     name: '5000원 할인 쿠폰',
     code: 'AMOUNT5000',
     discountType: 'amount',
-    discountValue: 5000
+    discountValue: 5000,
   },
   {
     name: '10% 할인 쿠폰',
     code: 'PERCENT10',
     discountType: 'percentage',
-    discountValue: 10
-  }
+    discountValue: 10,
+  },
 ];
 
 const TestAdminPage = () => {
   const [products, setProducts] = useState<Product[]>(mockProducts);
   const [coupons, setCoupons] = useState<Coupon[]>(mockCoupons);
 
-
   const handleProductUpdate = (updatedProduct: Product) => {
-    setProducts(prevProducts =>
-      prevProducts.map(p => p.id === updatedProduct.id ? updatedProduct : p)
+    setProducts((prevProducts) =>
+      prevProducts.map((p) => (p.id === updatedProduct.id ? updatedProduct : p))
     );
   };
 
   const handleProductAdd = (newProduct: Product) => {
-    setProducts(prevProducts => [...prevProducts, newProduct]);
+    setProducts((prevProducts) => [...prevProducts, newProduct]);
   };
 
   const handleCouponAdd = (newCoupon: Coupon) => {
-    setCoupons(prevCoupons => [...prevCoupons, newCoupon]);
+    setCoupons((prevCoupons) => [...prevCoupons, newCoupon]);
   };
 
   return (
@@ -76,18 +82,18 @@ const TestAdminPage = () => {
 };
 
 describe('basic > ', () => {
-
   describe('시나리오 테스트 > ', () => {
-
     test('장바구니 페이지 테스트 > ', async () => {
-
-      render(<CartPage products={mockProducts} coupons={mockCoupons}/>);
+      render(<CartPage products={mockProducts} coupons={mockCoupons} />);
       const product1 = screen.getByTestId('product-p1');
       const product2 = screen.getByTestId('product-p2');
       const product3 = screen.getByTestId('product-p3');
-      const addToCartButtonsAtProduct1 = within(product1).getByText('장바구니에 추가');
-      const addToCartButtonsAtProduct2 = within(product2).getByText('장바구니에 추가');
-      const addToCartButtonsAtProduct3 = within(product3).getByText('장바구니에 추가');
+      const addToCartButtonsAtProduct1 =
+        within(product1).getByText('장바구니에 추가');
+      const addToCartButtonsAtProduct2 =
+        within(product2).getByText('장바구니에 추가');
+      const addToCartButtonsAtProduct3 =
+        within(product3).getByText('장바구니에 추가');
 
       // 1. 상품 정보 표시
       expect(product1).toHaveTextContent('상품1');
@@ -99,7 +105,6 @@ describe('basic > ', () => {
       expect(product3).toHaveTextContent('상품3');
       expect(product3).toHaveTextContent('30,000원');
       expect(product3).toHaveTextContent('재고: 20개');
-
 
       // 2. 할인 정보 표시
       expect(screen.getByText('10개 이상: 10% 할인')).toBeInTheDocument();
@@ -159,17 +164,22 @@ describe('basic > ', () => {
     });
 
     test('관리자 페이지 테스트 > ', async () => {
-      render(<TestAdminPage/>);
-
+      render(<TestAdminPage />);
 
       const $product1 = screen.getByTestId('product-1');
 
       // 1. 새로운 상품 추가
       fireEvent.click(screen.getByText('새 상품 추가'));
 
-      fireEvent.change(screen.getByLabelText('상품명'), { target: { value: '상품4' } });
-      fireEvent.change(screen.getByLabelText('가격'), { target: { value: '15000' } });
-      fireEvent.change(screen.getByLabelText('재고'), { target: { value: '30' } });
+      fireEvent.change(screen.getByLabelText('상품명'), {
+        target: { value: '상품4' },
+      });
+      fireEvent.change(screen.getByLabelText('가격'), {
+        target: { value: '15000' },
+      });
+      fireEvent.change(screen.getByLabelText('재고'), {
+        target: { value: '30' },
+      });
 
       fireEvent.click(screen.getByText('추가'));
 
@@ -184,12 +194,17 @@ describe('basic > ', () => {
       fireEvent.click(within($product1).getByTestId('toggle-button'));
       fireEvent.click(within($product1).getByTestId('modify-button'));
 
-
       act(() => {
-        fireEvent.change(within($product1).getByDisplayValue('20'), { target: { value: '25' } });
-        fireEvent.change(within($product1).getByDisplayValue('10000'), { target: { value: '12000' } });
-        fireEvent.change(within($product1).getByDisplayValue('상품1'), { target: { value: '수정된 상품1' } });
-      })
+        fireEvent.change(within($product1).getByDisplayValue('20'), {
+          target: { value: '25' },
+        });
+        fireEvent.change(within($product1).getByDisplayValue('10000'), {
+          target: { value: '12000' },
+        });
+        fireEvent.change(within($product1).getByDisplayValue('상품1'), {
+          target: { value: '수정된 상품1' },
+        });
+      });
 
       fireEvent.click(within($product1).getByText('수정 완료'));
 
@@ -203,36 +218,57 @@ describe('basic > ', () => {
 
       // 할인 추가
       act(() => {
-        fireEvent.change(screen.getByPlaceholderText('수량'), { target: { value: '5' } });
-        fireEvent.change(screen.getByPlaceholderText('할인율 (%)'), { target: { value: '5' } });
-      })
+        fireEvent.change(screen.getByPlaceholderText('수량'), {
+          target: { value: '5' },
+        });
+        fireEvent.change(screen.getByPlaceholderText('할인율 (%)'), {
+          target: { value: '5' },
+        });
+      });
       fireEvent.click(screen.getByText('할인 추가'));
 
-      expect(screen.queryByText('5개 이상 구매 시 5% 할인')).toBeInTheDocument();
+      expect(
+        screen.queryByText('5개 이상 구매 시 5% 할인')
+      ).toBeInTheDocument();
 
       // 할인 삭제
       fireEvent.click(screen.getAllByText('삭제')[0]);
-      expect(screen.queryByText('10개 이상 구매 시 10% 할인')).not.toBeInTheDocument();
-      expect(screen.queryByText('5개 이상 구매 시 5% 할인')).toBeInTheDocument();
+      expect(
+        screen.queryByText('10개 이상 구매 시 10% 할인')
+      ).not.toBeInTheDocument();
+      expect(
+        screen.queryByText('5개 이상 구매 시 5% 할인')
+      ).toBeInTheDocument();
 
       fireEvent.click(screen.getAllByText('삭제')[0]);
-      expect(screen.queryByText('10개 이상 구매 시 10% 할인')).not.toBeInTheDocument();
-      expect(screen.queryByText('5개 이상 구매 시 5% 할인')).not.toBeInTheDocument();
+      expect(
+        screen.queryByText('10개 이상 구매 시 10% 할인')
+      ).not.toBeInTheDocument();
+      expect(
+        screen.queryByText('5개 이상 구매 시 5% 할인')
+      ).not.toBeInTheDocument();
 
       // 4. 쿠폰 추가
-      fireEvent.change(screen.getByPlaceholderText('쿠폰 이름'), { target: { value: '새 쿠폰' } });
-      fireEvent.change(screen.getByPlaceholderText('쿠폰 코드'), { target: { value: 'NEW10' } });
-      fireEvent.change(screen.getByRole('combobox'), { target: { value: 'percentage' } });
-      fireEvent.change(screen.getByPlaceholderText('할인 값'), { target: { value: '10' } });
+      fireEvent.change(screen.getByPlaceholderText('쿠폰 이름'), {
+        target: { value: '새 쿠폰' },
+      });
+      fireEvent.change(screen.getByPlaceholderText('쿠폰 코드'), {
+        target: { value: 'NEW10' },
+      });
+      fireEvent.change(screen.getByRole('combobox'), {
+        target: { value: 'percentage' },
+      });
+      fireEvent.change(screen.getByPlaceholderText('할인 값'), {
+        target: { value: '10' },
+      });
 
       fireEvent.click(screen.getByText('쿠폰 추가'));
 
       const $newCoupon = screen.getByTestId('coupon-3');
 
       expect($newCoupon).toHaveTextContent('새 쿠폰 (NEW10):10% 할인');
-    })
-  })
-
+    });
+  });
 
   describe('useProducts > ', () => {
     const initialProducts: Product[] = [
@@ -254,8 +290,8 @@ describe('basic > ', () => {
 
       expect(result.current.products[0]).toEqual({
         discounts: [],
-        id: "1",
-        name: "Updated Product",
+        id: '1',
+        name: 'Updated Product',
         price: 100,
         stock: 10,
       });
@@ -263,7 +299,13 @@ describe('basic > ', () => {
 
     test('새로운 제품을 추가할 수 있다.', () => {
       const { result } = renderHook(() => useProducts(initialProducts));
-      const newProduct: Product = { id: '2', name: 'New Product', price: 200, stock: 5, discounts: [] };
+      const newProduct: Product = {
+        id: '2',
+        name: 'New Product',
+        price: 200,
+        stock: 5,
+        discounts: [],
+      };
 
       act(() => {
         result.current.addProduct(newProduct);
@@ -282,7 +324,12 @@ describe('basic > ', () => {
 
     test('쿠폰을 추가할 수 있다', () => {
       const { result } = renderHook(() => useCoupons(mockCoupons));
-      const newCoupon: Coupon = { name: 'New Coupon', code: 'NEWCODE', discountType: 'amount', discountValue: 5000 };
+      const newCoupon: Coupon = {
+        name: 'New Coupon',
+        code: 'NEWCODE',
+        discountType: 'amount',
+        discountValue: 5000,
+      };
 
       act(() => {
         result.current.addCoupon(newCoupon);
@@ -301,8 +348,8 @@ describe('basic > ', () => {
       stock: 10,
       discounts: [
         { quantity: 2, rate: 0.1 },
-        { quantity: 5, rate: 0.2 }
-      ]
+        { quantity: 5, rate: 0.2 },
+      ],
     };
 
     describe('calculateItemTotal', () => {
@@ -332,7 +379,7 @@ describe('basic > ', () => {
     describe('calculateCartTotal', () => {
       const cart: CartItem[] = [
         { product: testProduct, quantity: 2 },
-        { product: {...testProduct, id: '2', price: 200}, quantity: 1 }
+        { product: { ...testProduct, id: '2', price: 200 }, quantity: 1 },
       ];
 
       test('쿠폰 없이 총액을 올바르게 계산해야 합니다.', () => {
@@ -343,14 +390,24 @@ describe('basic > ', () => {
       });
 
       test('금액쿠폰을 올바르게 적용해야 합니다.', () => {
-        const coupon: Coupon = { name: 'Test Coupon', code: 'TEST', discountType: 'amount', discountValue: 50 };
+        const coupon: Coupon = {
+          name: 'Test Coupon',
+          code: 'TEST',
+          discountType: 'amount',
+          discountValue: 50,
+        };
         const result = cartUtils.calculateCartTotal(cart, coupon);
         expect(result.totalAfterDiscount).toBe(330);
         expect(result.totalDiscount).toBe(70);
       });
 
       test('퍼센트 쿠폰을 올바르게 적용해야 합니다', () => {
-        const coupon: Coupon = { name: 'Test Coupon', code: 'TEST', discountType: 'percentage', discountValue: 10 };
+        const coupon: Coupon = {
+          name: 'Test Coupon',
+          code: 'TEST',
+          discountType: 'percentage',
+          discountValue: 10,
+        };
         const result = cartUtils.calculateCartTotal(cart, coupon);
         expect(result.totalAfterDiscount).toBe(342);
         expect(result.totalDiscount).toBe(58);
@@ -360,7 +417,7 @@ describe('basic > ', () => {
     describe('updateCartItemQuantity', () => {
       const cart: CartItem[] = [
         { product: testProduct, quantity: 2 },
-        { product: {...testProduct, id: '2'}, quantity: 1 }
+        { product: { ...testProduct, id: '2' }, quantity: 1 },
       ];
 
       test('수량을 올바르게 업데이트해야 합니다', () => {
@@ -383,8 +440,19 @@ describe('basic > ', () => {
   });
 
   describe('useCart > ', () => {
-    const testProduct: Product = { id: '1', name: 'Test Product', price: 100, stock: 10, discounts: [] };
-    const testCoupon: Coupon = { name: 'Test Coupon', code: 'TEST', discountType: 'percentage', discountValue: 10 };
+    const testProduct: Product = {
+      id: '1',
+      name: 'Test Product',
+      price: 100,
+      stock: 10,
+      discounts: [],
+    };
+    const testCoupon: Coupon = {
+      name: 'Test Coupon',
+      code: 'TEST',
+      discountType: 'percentage',
+      discountValue: 10,
+    };
 
     test('장바구니에 제품을 추가해야 합니다', () => {
       const { result } = renderHook(() => useCart());
@@ -394,7 +462,10 @@ describe('basic > ', () => {
       });
 
       expect(result.current.cart).toHaveLength(1);
-      expect(result.current.cart[0]).toEqual({ product: testProduct, quantity: 1 });
+      expect(result.current.cart[0]).toEqual({
+        product: testProduct,
+        quantity: 1,
+      });
     });
 
     test('장바구니에서 제품을 제거해야 합니다', () => {
@@ -444,5 +515,4 @@ describe('basic > ', () => {
       expect(total.totalDiscount).toBe(20);
     });
   });
-})
-
+});
